fix int overflow (thanks to Tobias (hawkeye) Pietzsch for spotting)

+++ mpicbg/ij/integral/BlockStatistics.java
@@ -112,7 +112,7 @@ final public class BlockStatistics
 			{
 				final int xMin = Math.max( -1, x - blockRadiusX - 1 );
 				final int xMax = Math.min( w, x + blockRadiusX );
-				final float scale = 1.0f / ( xMax - xMin ) / bh;
+				final double scale = 1.0 / ( xMax - xMin ) / bh;
 				fp.setf( row + x, ( float )( scale * sums.getDoubleSum( xMin, yMin, xMax, yMax ) ) );
 			}
 		}
@@ -146,7 +146,7 @@ final public class BlockStatistics
 			{
 				final int xMin = Math.max( -1, x - blockRadiusX - 1 );
 				final int xMax = Math.min( w, x + blockRadiusX );
-				final float scale = 1.0f / ( xMax - xMin ) / bh;
+				final double scale = 1.0 / ( xMax - xMin ) / bh;
 				final double sum = sums.getDoubleSum( xMin, yMin, xMax, yMax );
 				final double var = scale * ( sumsOfSquares.getDoubleSum( xMin, yMin, xMax, yMax ) - sum * sum * scale );
 				
@@ -192,9 +192,9 @@ final public class BlockStatistics
 //				final float scale = 1.0f / ( xMax - xMin ) / bh;
 //				final double sum = sums.getDoubleSum( xMin, yMin, xMax, yMax );
 //				final double var = scale * ( sumsOfSquares.getDoubleSum( xMin, yMin, xMax, yMax ) - sum * sum * scale );
-				final int bs = ( xMax - xMin ) * bh;
-				final float scale1 = 1.0f / ( bs - 1 );
-				final float scale2 = 1.0f / ( bs * bs - bs );
+				final long bs = ( xMax - xMin ) * bh;
+				final double scale1 = 1.0 / ( bs - 1 );
+				final double scale2 = 1.0 / ( bs * bs - bs );
 				final double sum = sums.getDoubleSum( xMin, yMin, xMax, yMax );
 				final double var = scale1 * sumsOfSquares.getDoubleSum( xMin, yMin, xMax, yMax ) - scale2 * sum * sum;
 				
@@ -238,9 +238,9 @@ final public class BlockStatistics
 			{
 				final int xMin = Math.max( -1, x - blockRadiusX - 1 );
 				final int xMax = Math.min( w, x + blockRadiusX );
-				final int bs = ( xMax - xMin ) * bh;
-				final float scale1 = 1.0f / ( bs - 1 );
-				final float scale2 = 1.0f / ( bs * bs - bs );
+				final long bs = ( xMax - xMin ) * bh;
+				final double scale1 = 1.0 / ( bs - 1 );
+				final double scale2 = 1.0 / ( bs * bs - bs );
 				final double sum = sums.getDoubleSum( xMin, yMin, xMax, yMax );
 				final double var = scale1 * sumsOfSquares.getDoubleSum( xMin, yMin, xMax, yMax ) - scale2 * sum * sum;
